requirements.txt
numpy==1.24.3
scipy==1.11.1
matplotlib==3.7.2
pyyaml==6.0.1
